renamed package name and removed off unused code and added authors details.
