--- grails-app/controllers/org/grails/plugins/csv/CsvTestController.groovy
@@ -1,16 +1,0 @@
-package org.grails.plugins.csv
-
-class CsvTestController {
-
-    def writeCsv = {
-        def rows = []
-        rows << [1,2]
-        rows << [3,4]
-        rows << [5,6]
-
-        renderCsv(rows: rows, filename: params.filename, attachment: params.attachment) {
-            x { it[0] }
-            y { it[1] }
-        }
-    }
-}
