--- grails-app/init/org/grails/plugins/csv/Application.groovy
@@ -1,10 +1,0 @@
-package org.grails.plugins.csv
-
-import grails.boot.GrailsApp
-import grails.boot.config.GrailsAutoConfiguration
-
-class Application extends GrailsAutoConfiguration {
-    static void main(String[] args) {
-        GrailsApp.run(Application, args)
-    }
-}
