--- resources/fileTest.csv
@@ -1,3 +1,0 @@
-a, "nested, commas", and a line "spanning
-a, few,
-lines", test
