--- resources/mapTest.csv
@@ -1,3 +1,0 @@
-col1,col2,col3
-"val1",val2,"val3"
-"10",20,"30,3"
